+++ lib/geolix/result/country.ex
@@ -21,7 +21,7 @@ defmodule Geolix.Result.Country do
       country:             Record.Country.from(data[:country], locale),
       registered_country:  Record.Country.from(data[:registered_country], locale),
       represented_country: Record.RepresentedCountry.from(data[:represented_country], locale),
-      traits:              %{}
+      traits:              Map.get(data, :traits, %{})
     }
   end
 end